requirements.txt
# Options Trading Bot - Dependencies
# ====================================

# Data & Computation
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Options Pricing
py_vollib>=1.0.1
mibian>=0.1.3

# Technical Analysis
ta>=0.10.2
pandas-ta>=0.3.14b

# Web Framework (Dashboard)
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0

# Broker APIs (Indian Markets)
kiteconnect>=5.0.0
smartapi-python>=1.3.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP & Async
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0

# Scheduling
apscheduler>=3.10.0
python-dateutil>=2.8.0

# Configuration & Logging
python-dotenv>=1.0.0
pyyaml>=6.0.0
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
click>=8.1.0
tabulate>=0.9.0
tqdm>=4.66.0
